requirements.txt
discord.py
aiogram>=3.0.0
aiohttp
